% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{crump_papaja}
\alias{crump_papaja}
\title{Set papaja yml}
\usage{
crump_papaja(papaja_file)
}
\arguments{
\item{papaja_file}{string, path and file name to a papaja.Rmd}
}
\value{
file, insert template yml to target document
}
\description{
Set papaja yml
}
